motion data graph now changes based on which radial chart is clicked
